requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.1
